separate the tests of the echo-post example
